refactor(new-ui): rename file
