+++ packages/new-polymath-ui/src/components/Form/index.ts
@@ -1,1 +1,1 @@
-export * from './Form';
+export * from './FormWrapper';
